define SerializeableRequest in the runtime crate, since the loader crate depends on std
