+++ experimental/uefi/loader/src/server.rs
@@ -18,9 +18,9 @@ use anyhow::bail;
 use bmrng::unbounded::UnboundedRequestSender;
 use futures::Future;
 use oak_remote_attestation_sessions::{SessionId, SESSION_ID_LENGTH};
+use runtime::SerializeableRequest;
 use std::net::SocketAddr;
 use tonic::{transport::Server, Request, Response, Status};
-use uefi_loader::SerializeableRequest;
 
 pub mod proto {
     tonic::include_proto!("oak.session.unary.v1");
