+++ experimental/uefi/runtime/Cargo.toml
@@ -9,5 +9,4 @@ license = "Apache-2.0"
 ciborium = { version = "*", default-features = false }
 ciborium-io = "*"
 oak_remote_attestation_sessions = { path = "../../../remote_attestation_sessions" }
-uefi-loader = { path = "../../../experimental/uefi/loader" }
 anyhow = { version = "*", default-features = false }
